test: sweep reward and upgrade test
